+++ src/helpers/chunkedDynamic.ts
@@ -2,15 +2,7 @@ import { clamp } from "./clamp";
 import { sleep } from "./sleep";
 
 /**
- * @description Run a series of (async) functions with dynamic chunk sizes
- * @param array - Array to chunk
- * @param initialChunkSize - Size of initial chunk
- * @param fn - Function to run on each chunk
- * @param options.idealChunkDuration - Ideal time to process each chunk, the chunk size will adjust to meet this duration
- * @param options.maxChunkSize - Maximum chunk size (default 200)
- * @param options.minChunkDuration - Minimum time to process each chunk (useful for rate limiting)
- * @param options.minChunkSize - Minimum chunk size (default 1)
- * @param options.sleepTimeMs - Time to sleep between each chunk
+ * @deprecated Use pMap/pAll/pReduce instead
  */
 export const chunkedDynamic = async <T>(
   array: T[],